Got player pieces to move.
Added a couple of instance variables to track current player, and all player positions, added loop to assign pngs to players, added function to get pixel positions of players for gui, added function to move piece the correct number of spaces (by move the picture the correct amount of pixels) and go to next players turn. Should work for up to 4 players in theory, but I only tested with 2.